backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
databases==0.6.2
Django==4.1.7
dnspython==2.3.0
docstring-parser==0.15
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
fastapi-pagination==0.11.4
geomet==0.2.1.post1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-model==0.0.5
MarkupSafe==2.1.2
mongoengine==0.26.0
motor==3.1.1
mypy-extensions==1.0.0
orjson==3.8.9
packaging==23.0
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pony==0.7.16
psycopg2==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pymongo==4.3.3
pypika-tortoise==0.1.6
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlakeyset==1.0.1679485999
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
sqlparse==0.4.3
starlette==0.26.1
targ==0.3.8
toml==0.10.2
tortoise-orm==0.19.3
typesystem==0.3.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==10.4
